+++ .gitignore
@@ -4,6 +4,33 @@ logs
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+pnpm-debug.log*
+lerna-debug.log*
+
+node_modules
+dist
+dist-ssr
+*.local
+
+# Editor directories and files
+.vscode/*
+!.vscode/extensions.json
+.idea
+.DS_Store
+*.suo
+*.ntvs*
+*.njsproj
+*.sln
+*.sw?
+
+#comparing
+
+# Logs
+logs
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
 lerna-debug.log*
 .pnpm-debug.log*
 
@@ -138,3 +165,11 @@ dist
 
 # macOS system files
 .DS_Store
+
+# Added from past shows feature branch
+.idea                  # JetBrains (WebStorm/PyCharm) settings
+*.suo                  # Visual Studio (Old) settings
+*.ntvs* # Node.js Tools for Visual Studio
+*.njsproj              # Node.js project file
+*.sln                  # Visual Studio Solution files
+.ipynb_checkpoints/    # Jupyter Notebook autosaves
