+++ packages/frontend/src/components/CardStack/.ipynb_checkpoints/CardStack-checkpoint.tsx
