+++ packages/frontend/src/components/CardStack/Card.css
