+++ packages/frontend/src/components/CardStack/Card.tsx
