+++ packages/frontend/src/components/CardStack/CardStack.tsx
