--- packages/frontend/src/pages/past_shows/.ipynb_checkpoints/PastShowsPage-checkpoint.css
@@ -1,9 +1,0 @@
-body {
-    overflow-x: hidden;
-}
-
-.hero {
-    background: #ffc6bf;
-    width: 100vw;
-    height: 90vh;
-}
