--- packages/frontend/src/pages/past_shows/components/.ipynb_checkpoints/Card-checkpoint.tsx
@@ -1,19 +1,0 @@
-import './Card.css'
-
-
-export interface PastShowCardProps {
-    name: string;
-}
-
-
-export default function Card({ info }: {info: PastShowCardProps}) {
-    
-    return(
-        
-        <div className="card">
-            <p>{info.name}</p>
-        </div>
-        
-    )
-        
-}
