--- packages/frontend/src/pages/past_shows/components/.ipynb_checkpoints/CardStack-checkpoint.css
@@ -1,10 +1,0 @@
-.card-stack{
-    width: 100%;
-    height: 100vh;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    /* Debug */
-    /* background: #fabfb4; */
-}
